(GP-2.3.1) Cleaning up Tester for Blob creation
Added a createblob method, which creats a file with contents and name equal to the hash code of the file passed into the method.
Cleaned up the tester:
    added the clearMakeFiles method which clears and makes the git, head, index, and objects folders
    added a makeBlobs method which makes blobs given Files
    added a checkBlobs method which checks if blobs for files exists
    added a createTestFiles method which creates testFiles and writes random content to them
    added a deleteTestFiles which deletes the testFiles.
Fixed a bug in 2.1 where index and HEAD were being created as folders, and fixed the delete files method in Git.java.